more match cleanup 2x
