+++ src/1-one/match/methods/match/steps/astrix.js
@@ -1,4 +1,4 @@
-import { greedyTo } from '../logic/greedy.js'
+import { greedyTo } from './logic/greedy.js'
 import { getGroup } from '../_lib.js'
 
 //support 'unspecific greedy' .* properly
