+++ src/1-one/match/methods/match/steps/logic/and-or.js
@@ -1,4 +1,4 @@
-import doesMatch from '../term/doesMatch.js'
+import doesMatch from '../../term/doesMatch.js'
 
 const isArray = function (arr) {
   return Object.prototype.toString.call(arr) === '[object Array]'
